requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.28.0
flask==2.3.3
joblib==1.3.2
pyyaml==6.0.1
scipy==1.11.1
jupyter==1.0.0
notebook==6.5.4
ipywidgets==8.1.1
tqdm==4.65.0
click==8.1.7
colorama==0.4.6
python-dotenv==1.0.0
gunicorn==21.2.0